Match Game is now completely functional
Updated board and candy script. For candy script: No able to move and switch a back with no match but if there is a match the candy will be destroyed.
The board script: The initial candies fall in at the start and now the board refills after candies are destroyed. Also added a state machine so that the player isn't able to move. the candy when it's still falling.